+++ resources/js/chat.js
@@ -5,7 +5,7 @@ import { v4 as uuidv4 } from 'uuid'
 
 const app = createApp(Chat)
 
-const config = window.botmanWidget
+const config = window.superbotmanWidget
 
 // add default page
 if (!config.pages?.length) {
@@ -71,7 +71,7 @@ const store = createStore({
             state.page = pageId
             if (pageId !== 'home') {
                 state.context = { ...state.context, pageId }
-                window.botmanWidget.context = { ...window.botmanWidget.context, pageId }
+                window.superbotmanWidget.context = { ...window.superbotmanWidget.context, pageId }
             }
             const page = state.config.pages.find(page => page.id === pageId)
             if (page?.pristine) {
@@ -110,12 +110,12 @@ const store = createStore({
         },
         context(state, data) {
             state.context = data
-            window.botmanWidget.context = data
+            window.superbotmanWidget.context = data
         },
         conversationId(state, { pageId, conversationId }) {
             state.conversationId[pageId] = conversationId
             state.context = { ...state.context, conversationId }
-            window.botmanWidget.context = { ...window.botmanWidget.context, conversationId }
+            window.superbotmanWidget.context = { ...window.superbotmanWidget.context, conversationId }
         },
         conversations(state, { pageId, conversations }) {
             state.conversations[pageId] = conversations
@@ -125,7 +125,7 @@ const store = createStore({
             state.conversationId[pageId] = null
             const { conversationId, ...rest } = state.context
             state.context = rest
-            window.botmanWidget.context = { ...rest }
+            window.superbotmanWidget.context = { ...rest }
             const page = state.config.pages.find(p => p.id === pageId)
             if (page) {
                 page.pristine = true